Add rotary bezel simulation code
Since Samsung Gear S2 has the rotary bezel and it cannot be emulated in
any 'native' way from the web browser, I had implemented a method that
uses 'j' and 'k' keys to rotate CCW and CW, respectively.

This method is only set up if the browser is not the one that executes
the application in the emulator or device, that is every web browser not
identified as 'Tizen'.