Remove all 'let' occurrences; fix rotation bug
I encountered problems trying to deploy the application in the emulator
by the CLI. The CLI just showed me the stacktrace of a Java exception,
so it was impossible to find the solution. Finally, I ran the
application  by the Tizen IDE and it showed me the errors that were
happening in the code...

And, listen to me, NEVER, I said, NEVER USE 'let' IN TIZEN. The
Javascript interpreter of the device has not implemented that keyword.

I fixed a small typo that makes the rotation method to fail, too.